Revive HelloWorldJobDemo and make it run at mvn:test time.
