+++ molgenis-jobs/src/test/resources/logback-test.xml
@@ -4,8 +4,12 @@
             <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
         </encoder>
     </appender>
+    <appender name="jobExecutionLogAppender" class="org.molgenis.data.jobs.model.JobExecutionLogAppender"/>
     <logger name="org.molgenis" level="INFO"/>
     <root level="WARN">
         <appender-ref ref="STDOUT"/>
     </root>
+    <logger name="org.molgenis.data.jobs.model.JobExecution" level="ALL">
+        <appender-ref ref="jobExecutionLogAppender"/>
+    </logger>
 </configuration>
